+++ lib/commands/services/project.ts
@@ -41,16 +41,25 @@ project
 
 project
   .command(`list-variables`)
-  .description(`Get a list of all project variables. These variables will be accessible in all Appwrite Functions at runtime.`)
+  .description(`Get a list of all project environment variables.`)
+  .option(`--queries [queries...]`, `Array of query strings generated using the Query class provided by the SDK. [Learn more about queries](https://appwrite.io/docs/queries). Maximum of 100 queries are allowed, each 4096 characters long. You may filter on the following attributes: key, resourceType, resourceId, secret`)
+  .option(
+    `--total [value]`,
+    `When set to false, the total count returned will be 0 and will not be calculated.`,
+    (value: string | undefined) =>
+      value === undefined ? true : parseBool(value),
+  )
   .action(
     actionRunner(
-      async () => parse(await (await getProjectClient()).listVariables()),
+      async ({ queries, total }) =>
+        parse(await (await getProjectClient()).listVariables(queries, total)),
     ),
   );
 
 project
   .command(`create-variable`)
-  .description(`Create a new project variable. This variable will be accessible in all Appwrite Functions at runtime.`)
+  .description(`Create a new project environment variable. These variables can be accessed by all functions and sites in the project.`)
+  .requiredOption(`--variable-id <variable-id>`, `Variable ID. Choose a custom ID or generate a random ID with \`ID.unique()\`. Valid chars are a-z, A-Z, 0-9, period, hyphen, and underscore. Can't start with a special char. Max length is 36 chars.`)
   .requiredOption(`--key <key>`, `Variable key. Max length: 255 chars.`)
   .requiredOption(`--value <value>`, `Variable value. Max length: 8192 chars.`)
   .option(
@@ -61,15 +70,15 @@ project
   )
   .action(
     actionRunner(
-      async ({ key, value, secret }) =>
-        parse(await (await getProjectClient()).createVariable(key, value, secret)),
+      async ({ variableId, key, value, secret }) =>
+        parse(await (await getProjectClient()).createVariable(variableId, key, value, secret)),
     ),
   );
 
 project
   .command(`get-variable`)
-  .description(`Get a project variable by its unique ID.`)
-  .requiredOption(`--variable-id <variable-id>`, `Variable unique ID.`)
+  .description(`Get a variable by its unique ID. `)
+  .requiredOption(`--variable-id <variable-id>`, `Variable ID.`)
   .action(
     actionRunner(
       async ({ variableId }) =>
@@ -79,9 +88,9 @@ project
 
 project
   .command(`update-variable`)
-  .description(`Update project variable by its unique ID. This variable will be accessible in all Appwrite Functions at runtime.`)
-  .requiredOption(`--variable-id <variable-id>`, `Variable unique ID.`)
-  .requiredOption(`--key <key>`, `Variable key. Max length: 255 chars.`)
+  .description(`Update variable by its unique ID.`)
+  .requiredOption(`--variable-id <variable-id>`, `Variable ID.`)
+  .option(`--key <key>`, `Variable key. Max length: 255 chars.`)
   .option(`--value <value>`, `Variable value. Max length: 8192 chars.`)
   .option(
     `--secret [value]`,
@@ -98,8 +107,8 @@ project
 
 project
   .command(`delete-variable`)
-  .description(`Delete a project variable by its unique ID. `)
-  .requiredOption(`--variable-id <variable-id>`, `Variable unique ID.`)
+  .description(`Delete a variable by its unique ID. `)
+  .requiredOption(`--variable-id <variable-id>`, `Variable ID.`)
   .action(
     actionRunner(
       async ({ variableId }) =>